Example function for creating a mapping from a virtual page to a physical frame
This maps some chosen virtual page to the physical frame 0xb8000, where
the VGA buffer resides, just as an example.

It includes two implementations of FrameAllocator; one which can
allocate new frames (for the creation of new page tables as required)
and one which cannot, so it only works for mapping pages in existing
page tables.

Source: https://os.phil-opp.com/paging-implementation/