Initialise a heap
Here we create a heap section in virtual memory by requesting available
physical frames and mapping each page in the heap to them.

This still doesn't allow Box to work as we don't have an allocator that
can provide space from the heap for the program to use.

Source: https://os.phil-opp.com/heap-allocation/