Implement a fixed-size block allocator
This is a cache over the `linked_list_allocator` fallback. Block
requests are rounded up to a power of two and initially allocated from
the fallback. On deallocation, a block is put in a list for fast reuse.

Just like in the linked list allocator, the list nodes are stored in the
free block space itself. We're effectively indexing blocks by size, so
rather than having to search for a block that's big enough, and possibly
split it up, we just use the head of the list for the requested size.

Source: https://os.phil-opp.com/allocator-designs/