Add Waker support to the executor
This replaces SimpleExecutor with Executor, a better implementation that
actually makes use of wakers to re-schedule tasks, rather than polling
all tasks in a loop. It means that if there are no tasks making process,
we can run HLT and put the CPU to sleep.

Source: https://os.phil-opp.com/async-await/